updating example
