+++ example/lib/ui/auth/check.dart
@@ -1,0 +1,24 @@
+import 'package:fb_auth/fb_auth.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+import '../auth/screen.dart';
+import '../home/screen.dart';
+
+class AuthCheck extends StatelessWidget {
+  const AuthCheck({
+    Key key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocBuilder<AuthBloc, AuthState>(
+      builder: (context, state) {
+        if (state is LoggedInState) {
+          return HomeScreen();
+        }
+        return LoginScreen();
+      },
+    );
+  }
+}
